lr should be a float

+++ src/yogo/argparsers.py
@@ -46,7 +46,7 @@ def train_parser(parser=None):
         "--batch-size", type=uint, help="batch size for training", default=None
     )
     parser.add_argument(
-        "--lr", type=uint, help="learning rate for training", default=None
+        "--lr", type=float, help="learning rate for training", default=None
     )
     parser.add_argument(
         "--epochs", type=uint, help="number of epochs to train", default=None
